fix: Current Stake Logic and UX
Improve the design of the current stakes layout.

Also fixes the current logic by refactoring the separate lists of active
stakes and active unstakes into a list of positions, where a position
represents a validator and collects its list of active stakes and
unstakes.

It turns out the validator response does in fact return the recent
uptime of a validator, despite the missing property in the SDK