fix: Stake List was using an unpredicatable key
The list of current positions displayed on the stake page was depending
on the index of the stake in the sorted list as a key.  Over time, the
index of a position can change on the sorted list as a user interacts
with the positions and as the values of the positions change from
polling.

This commit fixes these inconsistencies by depending on the address of
the valdiator as a key, extending the `Position` type by adding the
address as a string.

This PR also fixes an out of date `CopyToClipboard` call that was
sending a `text` prop rather then the updated `address` prop.