fix: Ledger Immediately Unavailalble after creating a new Wallet (#364)
This commit fixes a problem with the SDK Wallet subscription receiving
multiple wallet subjects while creating a new wallet.  This collection
of wallet instances then all forwarded the input for to the ledger
device when deriving a new HW account which resulted in a race condition
and UI Crash.

The work contained in this commit steps away from the RXJS subscriptions
by using `firstValueFrom` to derive promises for both the active wallet
and the first result of `deriveHWAddress` on the derived wallet, thus
only sending a single message to the ledger.