fix: Network Change Reset (#512)
When a user chooses to connect to a different node in the settings, the
app should restart and they should be required to log back in.

Currently the reload was occurring but they were not being redirected to
the home page.  This was a regression from the refactoring of the url
params.